[ISPC] Fix #3 Accessing xyz in a float4 isn't supported.
Although the original bug was not reproduced, a very similar issue was. It was found that a swizzle could be generated, as an optimisation, when building a composite combiner which was then invalid for ispc.

This has been disabled via a new backend compiler option :

    backend.supports_native_swizzle

which defaults to true, but is set to false for ISPC.