[ISPC] Fix #5 Add basic support for group shared memory.
This is now supported by creating a uniform variable in the dispatch framework and passing it down into the kernel. This will be shared by all instances of the kernel within the workgroup.

Shared memory arrays are passed as pointers and non array types are passed as references.

Further work may be required for this support.