[ISPC] Fix #4 Add support for varying constant literals
There is often the situation where shader code is written such that a constant index is used to dereference a uniform array and store a varying variable into it. ie.

    if (varyingID == 0)
        uniformArray[0] = varyingResult

If '0' is a uniform literal, then LHS is uniform and RHS is varying, giving a compiler error.

Hence, this needs expressing as :

    if (varyingID == 0)
        uniformArray[(varying int)0] = varyingResult

Allowing both LHS and RHS to be varyings.

This is implemented by detecting constant lookups in access chains and creating a new varying version of them. When the access chain is due to be written out, if it is varying then the varying constant is used instead of the uniform constant.