[ISPC] Fix #6. Group barriers and syncs now throw an exception unless explicitly ignored via a command line argument.
ISPC behaves differently to DX/Vulkan/CL in that a workgroup does not map to a number of individual execution threads in the hardware which may require synchronisation. ISPC iterates over the NDRange in chunks of SIMD width work group items. As such, it has no concept of barriers.

The ISPC backend will now throw an exception if it finds a group barrier or sync, but this can be avoided if the user supplies the following command line option :

    --ispc-ignore-group-barriers